Updating build scripts to make them more modular with options

+++ tools/Build-Nuget.ps1
@@ -1,47 +1,25 @@
 param(
     [switch]$upload = $false,
     [switch]$fast = $false,
-    [switch]$skipTest = $false,
-    [string]$key = ''
+    [switch]$noTest = $false,
+    [switch]$noBuild = $false,
+    [string]$key = '',
+    [string]$saveKey = ''
 )
 
 function Get-LineNumber { return $MyInvocation.ScriptLineNumber }
 function Get-ScriptName { return $MyInvocation.ScriptName }
 
 # In case we only want to build a subset of the package
-$buildWindows    = $true
-$buildWindowsUWP = $true
-$buildLinux      = $true
-$buildAndroid    = $true
+$buildWindows    = $true -and -not $noBuild
+$buildWindowsUWP = $true -and -not $noBuild
+$buildLinux      = $true -and -not $noBuild
+$buildAndroid    = $true -and -not $noBuild
+
+$clean = $true -and -not $noBuild
 
 ###########################################
 ## Functions                             ##
-###########################################
-
-function Replace-In-File {
-    param($file, $text, $with)
-
-    ((Get-Content -path $file) -replace $text,$with) | Set-Content -path $file
-}
-
-###########################################
-
-function Get-Key {
-    if ($key -ne '') {
-        Set-Content -path '.nugetkey' -value $key.Trim()
-    } elseif (Test-Path '.nugetkey') {
-        $key = Get-Content -path '.nugetkey' -Raw
-    } else {
-        $key = Read-Host "Please enter the NuGet API key, or n to cancel"
-        if ($key -eq 'n') {
-            $key = ''
-        } else {
-            Set-Content -path '.nugetkey' -value $key.Trim()
-        }
-    }
-    return $key.Trim()
-}
-
 ###########################################
 
 function Build-Preset {
@@ -59,56 +37,18 @@ function Build-Preset {
 }
 
 ###########################################
-
-function Build-Sizes {
-    $size_x64        = (Get-Item "bin/distribute/bin/Win32/x64/Release/StereoKitC.dll").length
-    $size_x64_linux  = (Get-Item "bin/distribute/bin/Linux/x64/release/libStereoKitC.so").length
-    $size_x64_uwp    = (Get-Item "bin/distribute/bin/UWP/x64/Release/StereoKitC.dll").length
-    $size_arm64      = (Get-Item "bin/distribute/bin/Android/arm64-v8a/release/libStereoKitC.so").length
-    $size_arm64_linux= (Get-Item "bin/distribute/bin/Linux/arm64/release/libStereoKitC.so").length
-    $size_arm64_uwp  = (Get-Item "bin/distribute/bin/UWP/ARM64/Release/StereoKitC.dll").length
-    $size_arm_uwp    = (Get-Item "bin/distribute/bin/UWP/ARM/Release/StereoKitC.dll").length
-
-    $text = (@"
-## Build Sizes:
-
-| Platform | Arch  | Size, kb | Size, bytes |
-| -------- | ----- | -------- | ----------- |
-| Win32    | x64   | {0,8:N0} | {1,11:N0} |
-| UWP      | x64   | {2,8:N0} | {3,11:N0} |
-| UWP      | ARM64 | {4,8:N0} | {5,11:N0} |
-| UWP      | ARM   | {6,8:N0} | {7,11:N0} |
-| Linux    | x64   | {8,8:N0} | {9,11:N0} |
-| Linux    | ARM64 | {10,8:N0} | {11,11:N0} |
-| Android  | ARM64 | {12,8:N0} | {13,11:N0} |
-"@ -f ([math]::Round($size_x64        /1kb), $size_x64,
-       [math]::Round($size_x64_uwp    /1kb), $size_x64_uwp,
-       [math]::Round($size_arm64_uwp  /1kb), $size_arm64_uwp,
-       [math]::Round($size_arm_uwp    /1kb), $size_arm_uwp,
-       [math]::Round($size_x64_linux  /1kb), $size_x64_linux,
-       [math]::Round($size_arm64_linux/1kb), $size_arm64_linux,
-       [math]::Round($size_arm64      /1kb), $size_arm64
-       ))
-
-    return $text
-}
-
-###########################################
 ## Main                                  ##
 ###########################################
 
 # Notify about our upload flag status 
-if ($fast -eq $true -and $upload -eq $true) {
+if ($fast -and $upload) {
     Write-Host "Let's not upload a fast build, just in case! Try again without the fast flag :)" -ForegroundColor yellow
     exit
 }
 Write-Host 'Building... ' -NoNewline
-if ($upload -eq $false) {
-    Write-Host 'local only.' -ForegroundColor White
-} else {
-    Write-Host 'AND UPLOADING!' -ForegroundColor Green
-}
-if ($fast -eq $true) {
+if ($upload) { Write-Host 'AND UPLOADING!' -ForegroundColor Green }
+else         { Write-Host 'local only.'    -ForegroundColor White }
+if ($fast) {
     Write-Host 'Making a "fast" build, incremental build issues may be present.'
 }
 
@@ -118,21 +58,9 @@ Push-Location -Path "$PSScriptRoot\.."
 
 #### Update Version #######################
 
-# Print version, so we know we're building the right version right away
-$fileData = Get-Content -path 'StereoKitC\stereokit.h' -Raw;
-$fileData -match '#define SK_VERSION_MAJOR\s+(?<ver>\d+)' | Out-Null
-$major = $Matches.ver
-$fileData -match '#define SK_VERSION_MINOR\s+(?<ver>\d+)' | Out-Null
-$minor = $Matches.ver
-$fileData -match '#define SK_VERSION_PATCH\s+(?<ver>\d+)' | Out-Null
-$patch = $Matches.ver
-$fileData -match '#define SK_VERSION_PRERELEASE\s+(?<ver>\d+)' | Out-Null
-$pre = $Matches.ver
-
-$version = "$major.$minor.$patch"
-if ($pre -ne 0) {
-    $version = "$version-preview.$pre"
-}
+$version = & $PSScriptRoot\Get-Version.ps1
+# Ensure all versions match the source of truth
+& $PSScriptRoot\Set-Version.ps1 -major $version.major -minor $version.minor -patch $version.patch -pre $version.pre
 
 # Notify of build, and output the version
 Write-Host @"
@@ -143,31 +71,30 @@ Write-Host @"
   ____) | ||  __/ | |  __/ ( ) |   \| | |_ 
  |_____/ \__\___|_|  \___|\___/|_|\_\_|\__| 
 "@ -NoNewline -ForegroundColor White
-Write-Host "v$version`n" -ForegroundColor Cyan
-
-# Ensure the version string for the package matches the StereoKit version
-Replace-In-File -file 'StereoKit\StereoKit.csproj' -text '<Version>(.*)</Version>' -with "<Version>$version</Version>"
-Replace-In-File -file 'CMakeLists.txt' -text 'StereoKit VERSION "(.*)"' -with "StereoKit VERSION `"$major.$minor.$patch`""
+Write-Host "v$($version.str)`n" -ForegroundColor Cyan
 
 #### Clean Project ########################
 
 # Clean out the old files, do a full build
-Write-Host 'Cleaning old files...'
-if (Test-Path 'bin\distribute') {
-    Remove-Item 'bin\distribute' -Recurse
-}
-if ($fast -eq $false) {
-    if (Test-Path 'bin\intermediate\cmake') {
-        Remove-Item 'bin\intermediate\cmake' -Recurse -Force
+if ($clean) {
+    Write-Host 'Cleaning old files...'
+    if (Test-Path 'bin\distribute') {
+        Remove-Item 'bin\distribute' -Recurse
     }
+    if ($fast -eq $false) {
+        if (Test-Path 'bin\intermediate\cmake') {
+            Remove-Item 'bin\intermediate\cmake' -Recurse -Force
+        }
+    }
+    Write-Host 'Cleaned'
+} else {
+    Write-Host 'Skipping clean'
 }
-Write-Host 'Cleaned'
 
 #### Build Windows ########################
 
 if ($buildWindows) {
     Write-Host @"
-
 __      __          _               
 \ \    / ( )_ _  __| |_____ __ _____
  \ \/\/ /| | ' \/ _  / _ \ V  V (_-<
@@ -186,7 +113,6 @@ __      __          _
 #### Build Linux ##########################
 if ($buildLinux) {
     Write-Host @"
-
   _                   
  | |  (_)_ _ _  ___ __
  | |__| | ' \ || \ \ /
@@ -202,7 +128,6 @@ if ($buildLinux) {
 
 if ($buildAndroid) {
     Write-Host @"
-
     _           _              _ 
    /_\  _ _  __| |_ _ ___( )__| |
   / _ \| ' \/ _' | '_/ _ \ / _' |
@@ -217,8 +142,14 @@ if ($buildAndroid) {
 #### Run tests ########################
 
 # Run tests!
-if ($fast -eq $false -and $skipTest -eq $false) {
-    Run-Tests
+if ($fast -eq $false -and $noTest -eq $false) {
+    & $PSScriptRoot/Run-Tests.ps1
+    if ($LASTEXITCODE -ne 0) {
+        Write-Host '--- Tests failed! Stopping build! ---' -ForegroundColor red
+        Pop-Location
+        exit
+    }
+    Write-Host 'Tests passed!' -ForegroundColor green
 } else {
     Write-Host "`nSkipping tests for fast build!" -ForegroundColor yellow
 }
@@ -226,7 +157,6 @@ if ($fast -eq $false -and $skipTest -eq $false) {
 #### Assemble NuGet Package ###############
 
 Write-Host @"
-
   _  _       ___     _   
  | \| |_  _ / __|___| |_ 
  |    | || | (_ / -_)  _|
@@ -236,11 +166,14 @@ Write-Host @"
 
 Write-Host "--- Beginning build: NuGet package ---" -ForegroundColor green
 # Turn on NuGet package generation, build, then turn it off again
-$packageOff = '<GeneratePackageOnBuild>false</GeneratePackageOnBuild>'
-$packageOn  = '<GeneratePackageOnBuild>true</GeneratePackageOnBuild>'
-Replace-In-File -file 'StereoKit\StereoKit.csproj' -text $packageOff -with $packageOn
-& dotnet build StereoKit\StereoKit.csproj
-Replace-In-File -file 'StereoKit\StereoKit.csproj' -text $packageOn -with $packageOff
+$packageOff  = '<GeneratePackageOnBuild>false</GeneratePackageOnBuild>'
+$packageOn   = '<GeneratePackageOnBuild>true</GeneratePackageOnBuild>'
+$packageFile = "$PSScriptRoot/../StereoKit/StereoKit.csproj"
+$partialBuild= '-p:SKIgnoreMissingBinaries=true'
+if ($upload) {$partialBuild = ''}
+[System.IO.File]::WriteAllText($packageFile, ((Get-Content -Path $packageFile -raw) -replace $packageOff, $packageOn))
+& dotnet build StereoKit\StereoKit.csproj $partialBuild -c Release
+[System.IO.File]::WriteAllText($packageFile, ((Get-Content -Path $packageFile -raw) -replace $packageOn, $packageOff))
 if ($LASTEXITCODE -ne 0) {
     Write-Host '--- NuGet build failed! Stopping build! ---' -ForegroundColor red
     Pop-Location
@@ -250,17 +183,31 @@ Write-Host "--- Finished building: NuGet package ---"-ForegroundColor green
 
 #### Create Build Info File ###############
 
-$build_size = Build-Sizes
-$build_info = "# StereoKit v$version Build Information
-
-$build_size"
-Set-Content -path 'tools\BuildInfo.md' -value $build_info
-Write-Host $build_info
+& $PSScriptRoot\Make-Bloat-Report.ps1
 
 #### Upload NuGet Package #################
 
 if ($upload) {
-    $key = Get-Key
+    # Figure out what key to use when uploading
+    if ($saveKey -ne '') {
+        # User wants us to cache this key in a file for reuse later
+        $key = $saveKey
+        Set-Content -path '.nugetkey' -value $key.Trim()
+    } elseif ($key -ne '') {
+        # User provided a key as an argument
+    } elseif (Test-Path '.nugetkey') {
+        # No user key, check for a cached key
+        $key = Get-Content -path '.nugetkey' -Raw
+    } else {
+        # No key, ask the user for one
+        $key = Read-Host "Please enter the NuGet API key, or n to cancel"
+        if ($key -eq 'n') {
+            $key = ''
+        }
+    }
+    $key = $key.Trim()
+
+    # Upload with our key!
     if ($key -ne '') {
         & dotnet nuget push "bin\StereoKit.$version.nupkg" -k $key -s https://api.nuget.org/v3/index.json
     } else {
